PPT_generator/requirements.txt
streamlit
python-pptx
llama-cpp-python
python-dotenv